Configure API_HOST_URL from test class and use it for mocked server tests
