+++ VideoLocker/default_config/config.yaml
@@ -1,4 +1,4 @@
-API_HOST_URL: 'http://localhost:8000/mock/'
+API_HOST_URL: 'http://localhost:8000'
 ENVIRONMENT_DISPLAY_NAME: 'Localhost'
 FEEDBACK_EMAIL_ADDRESS: 'support@example.com'
 OAUTH_CLIENT_ID: ''
