+++ VideoLocker/src/main/java/org/edx/mobile/module/serverapi/IRequest.java
@@ -20,4 +20,12 @@ public interface IRequest {
     void addHeader(String key, String value);
     Bundle getHeaders();
     void setHeaders(Bundle headers);
+
+    public static final class Method {
+        public static final String GET      = "GET";
+        public static final String POST     = "POST";
+        public static final String PUT      = "PUT";
+        public static final String DELETE   = "DELETE";
+        public static final String PATCH    = "PATCH";
+    }
 }
