+++ VideoLocker/src/main/java/org/edx/mobile/module/serverapi/IRequestImpl.java
@@ -67,12 +67,4 @@ class IRequestImpl implements IRequest {
     public void setHeaders(Bundle headers) {
         this.headers = headers;
     }
-
-    public static final class Method {
-        public static final String GET      = "GET";
-        public static final String POST     = "POST";
-        public static final String PUT      = "PUT";
-        public static final String DELETE   = "DELETE";
-        public static final String PATCH    = "PATCH";
-    }
 }
