+++ VideoLocker/src/main/java/org/edx/mobile/module/serverapi/parser/IParser.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface IParser {
 
-	public <T> T parseObject(String data, Class<T> cls) throws Exception;
+	public <T> T parseObject(String jsonObject, Class<T> cls) throws Exception;
     public <T> T parseObject(Reader reader, Class<T> cls) throws Exception;
-	public <T> List<T> parseList(String data, Class<T> cls) throws Exception;
+	public <T> List<T> parseList(String jsonArray, Class<T> cls) throws Exception;
 }
