+++ VideoLocker/src/main/java/org/edx/mobile/module/serverapi/parser/ParserFactory.java
@@ -2,16 +2,11 @@ package org.edx.mobile.module.serverapi.parser;
 
 public class ParserFactory {
 	
-    private static IParser instance;
-
     /**
-     * Returns singleton instance of {@link IParser} interface.
+     * Returns new instance of {@link IParser} interface.
      * @return
      */
     public static IParser getInstance() {
-        if (instance == null) {
-            instance = new GsonParser();
-        }
-        return instance;
+        return new GsonParser();
     }
 }
